day 6/requirements.txt
streamlit
langchain
PyPDF2
faiss-cpu
python-dotenv
google-generativeai
langchain-openai